DepartementService modified

+++ src/main/java/com/nerdtic/humanos/repositories/ContratRepository.java
@@ -3,6 +3,6 @@ package com.nerdtic.humanos.repositories;
 import com.nerdtic.humanos.entities.Contrat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContratRepository extends JpaRepository<Contrat, Integer> {
+public interface ContratRepository extends JpaRepository<Contrat, Long> {
 
 }
