+++ src/main/java/com/nerdtic/humanos/repositories/DepartementRepository.java
@@ -3,6 +3,6 @@ package com.nerdtic.humanos.repositories;
 import com.nerdtic.humanos.entities.Departement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartementRepository extends JpaRepository<Departement, Integer> {
+public interface DepartementRepository extends JpaRepository<Departement, Long> {
 
 }
