+++ src/main/java/com/nerdtic/humanos/repositories/FormationRepository.java
@@ -3,5 +3,5 @@ package com.nerdtic.humanos.repositories;
 import com.nerdtic.humanos.entities.Formation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FormationRepository extends JpaRepository<Formation, Integer> {
+public interface FormationRepository extends JpaRepository<Formation, Long> {
 }
